+++ mcp_tools/param_discovery/arjun.py
@@ -38,3 +38,38 @@ def register_arjun_tool(mcp, hexstrike_client, logger):
         else:
             logger.error(f"❌ Arjun parameter discovery failed for {url}")
         return result
+    
+    @mcp.tool()
+    def arjun_scan(url: str, method: str = "GET", data: str = "", headers: str = "", timeout: str = "", output_file: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Arjun for parameter discovery with enhanced logging.
+
+        Args:
+            url: Target URL
+            method: HTTP method (GET, POST, etc.)
+            data: POST data for testing
+            headers: Custom headers
+            timeout: Request timeout
+            output_file: Output file path
+            additional_args: Additional Arjun arguments
+
+        Returns:
+            Parameter discovery results
+        """
+        payload = {
+            "url": url,
+            "method": method,
+            "data": data,
+            "headers": headers,
+            "timeout": timeout,
+            "output_file": output_file,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting Arjun parameter discovery: {url}")
+        result = hexstrike_client.safe_post("api/tools/arjun", payload)
+        if result.get("success"):
+            logger.info(f"✅ Arjun completed for {url}")
+        else:
+            logger.error(f"❌ Arjun failed for {url}")
+        return result
+
