+++ mcp_tools/web_app_security/enhanced_web_app_security.py
@@ -5,78 +5,6 @@ from typing import Dict, Any, Optional
 def register_enhanced_web_app_security_tools(mcp, hexstrike_client, logger, HexStrikeColors):
 
     @mcp.tool()
-    def zap_scan(target: str = "", scan_type: str = "baseline", api_key: str = "", daemon: bool = False, port: str = "8090", host: str = "0.0.0.0", format_type: str = "xml", output_file: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute OWASP ZAP with enhanced logging.
-
-        Args:
-            target: Target URL
-            scan_type: Type of scan (baseline, full, api)
-            api_key: ZAP API key
-            daemon: Run in daemon mode
-            port: Port for ZAP daemon
-            host: Host for ZAP daemon
-            format_type: Output format (xml, json, html)
-            output_file: Output file path
-            additional_args: Additional ZAP arguments
-
-        Returns:
-            ZAP scan results
-        """
-        data = {
-            "target": target,
-            "scan_type": scan_type,
-            "api_key": api_key,
-            "daemon": daemon,
-            "port": port,
-            "host": host,
-            "format": format_type,
-            "output_file": output_file,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting ZAP scan: {target}")
-        result = hexstrike_client.safe_post("api/tools/zap", data)
-        if result.get("success"):
-            logger.info(f"✅ ZAP scan completed for {target}")
-        else:
-            logger.error(f"❌ ZAP scan failed for {target}")
-        return result
-
-    @mcp.tool()
-    def arjun_scan(url: str, method: str = "GET", data: str = "", headers: str = "", timeout: str = "", output_file: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Arjun for parameter discovery with enhanced logging.
-
-        Args:
-            url: Target URL
-            method: HTTP method (GET, POST, etc.)
-            data: POST data for testing
-            headers: Custom headers
-            timeout: Request timeout
-            output_file: Output file path
-            additional_args: Additional Arjun arguments
-
-        Returns:
-            Parameter discovery results
-        """
-        payload = {
-            "url": url,
-            "method": method,
-            "data": data,
-            "headers": headers,
-            "timeout": timeout,
-            "output_file": output_file,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Arjun parameter discovery: {url}")
-        result = hexstrike_client.safe_post("api/tools/arjun", payload)
-        if result.get("success"):
-            logger.info(f"✅ Arjun completed for {url}")
-        else:
-            logger.error(f"❌ Arjun failed for {url}")
-        return result
-
-    @mcp.tool()
     def wafw00f_scan(target: str, additional_args: str = "") -> Dict[str, Any]:
         """
         Execute wafw00f to identify and fingerprint WAF products with enhanced logging.
